aider: Make card components resize immediately upon overflow by updating media queries in SCSS files for `ServiceCardL` and `ServiceCardR`.
# Aider chat conversation:

USER: i want the card components to resize as soon as there is any kind of overflow. i dont want to wait till the viewport width is less than 720

ASSISTANT: To make the card components resize as soon as there is any kind of overflow without waiting for the viewport width to be less than 720px, we need to modify the media queries in the SCSS files for both `ServiceCardL` and `ServiceCardR` components. We will change the media query to trigger on `max-width: 100%` to ensure that the resizing happens immediately upon overflow.

Here are the diffs for the `service-card-l.module.scss` and `service-card-r.module.scss` files:

For `service-card-l.module.scss`:

```diff
--- src/components/service-card-l/service-card-l.module.scss
+++ src/components/service-card-l/service-card-l.module.scss
@@ -35,7 +35,7 @@
     max-width: 100%;
     max-height: 100%;

-    @media (max-width: $mobile-width) {
+    @media (max-width: 100%) {
         height: auto;
         width: auto; /* Allow the card to fit the mobile screen */
     }
```

For `service-card-r.module.scss`:

```diff
--- src/components/service-card-r/service-card-r.module.scss
+++ src/components/service-card-r/service-card-r.module.scss
@@ -35,7 +35,7 @@
     max-width: 100%;
     max-height: 100%;

-    @media (max-width: $mobile-width) {
+    @media (max-width: 100%) {
         flex-direction: column;
         height: auto;
     }
@@ -44,7 +44,7 @@
     max-width: 100%;
     max-height: 100%;

-    @media (max-width: $mobile-width) {
+    @media (max-width: 100%) {
         width: 100%;
         height: auto;
         padding: 0;
```

These changes will make the card components responsive to any overflow by resizing them immediately.