aider: Set the background color to glossy black for the main container, header, and footer, and make the header and footer backgrounds transparent.
# Aider chat conversation:

USER: what i meant was that the homepage background of glossy black should extend into the areas of header and footer as well. I have rounded edged on both header and footer. The area outisde the header's round edge and outside the homepage component appears white. similarly in the case of footer. do you know what I mean?

ASSISTANT: To achieve the effect of the glossy black background extending into the areas of the header and footer, we need to set the background of the main container that holds the homepage, header, and footer to the glossy black color. This way, the rounded edges of the header and footer will reveal the glossy black background of the main container instead of the default white.

Here's what you need to do:

1. Set the background color of the main container (which might be the `body` or a main `div` that wraps around the entire content including header, footer, and homepage) to glossy black.
2. Ensure that the header and footer have transparent backgrounds so that the glossy black from the main container shows through.

Since I don't have the exact structure of your main container, I'll assume it's the `body` element. If it's not, you'll need to apply the background color to the correct container element.

Here are the diffs for the changes:

```diff
--- src\styles\theme.module.scss
+++ src\styles\theme.module.scss
@@ ... @@
 // colors
 // Colors
-$white: #FFF;
+$white: #1c1c1c; // Change the white color variable to glossy black
 $black: #000;
 $charcoal: #282B2D;
 $snow: #F5F5F5;
```

```diff
--- src\components\header\header.module.scss
+++ src\components\header\header.module.scss
@@ ... @@
 .root {
     display: flex;
     flex-direction: row;
     padding: 0 3em; // Unify padding for left and right
     justify-content: space-between;
     align-items: center;
     height: 4em;
     font: $button-text;
     font: $button-text;
-    background-color: $white; // Revert to original white background color
+    background-color: transparent; // Make the background transparent
     top: 0;
     left: 0;
     width: 100%;
     z-index: 1000; /* Ensure the header is above other content */
```

```diff
--- src\components\footer\footer.module.scss
+++ src\components\footer\footer.module.scss
@@ ... @@
 .root {
     display: flex;
     justify-content: space-between;
     width: 100%;
     padding-right: 3.5em;
     padding-left: 3em;
     padding-top: 2em;
     padding-bottom: 1.2em;
     opacity: 0.9;
-    background-image: linear-gradient(white, white); // Revert to original white background gradient
+    background-color: transparent; // Make the background transparent
     background-position: 0%;
     background-size: auto;
     background-repeat: repeat;
```

Make sure to apply the glossy black background to the main container of your application. If it's not the `body` element, you'll need to adjust the diffs accordingly.